+++ buildpacks/jvm/CHANGELOG.md
@@ -3,6 +3,28 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+* Switch to BSD 3-Clause License
+
+## [0.1.11] 2021/10/28
+### Changed
+* Default version for **OpenJDK 7** is now `1.7.0_322`
+* Default version for **OpenJDK 17** is now `17.0.1`
+
+## [0.1.10] 2021/10/27
+
+## [0.1.9] 2021/10/19
+### Changed
+* Default version for **OpenJDK 8** is now `1.8.0_312`
+* Default version for **OpenJDK 11** is now `11.0.13`
+* Default version for **OpenJDK 13** is now `13.0.9`
+* Default version for **OpenJDK 15** is now `15.0.5`
+
+## [0.1.8] 2021/09/15
+### Added
+* Support for Java 17
+
+### Fixed
+* Updated GPG public key
 
 ## [0.1.7] 2021/07/28
 ### Changed