+++ meta-buildpacks/java-function/CHANGELOG.md
@@ -3,6 +3,32 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+* Switch to BSD 3-Clause License
+
+## [0.3.24] 2021/10/28
+* Upgraded `heroku/jvm` to `0.1.11`
+* Upgraded `heroku/jvm` to `0.1.10`
+
+## [0.3.23] 2021/10/19
+* Upgraded `heroku/jvm-function-invoker` to `0.5.5`
+* Upgraded `heroku/jvm` to `0.1.9`
+
+
+## [0.3.22] 2021/09/30
+* Upgraded `heroku/jvm-function-invoker` to `0.5.4`
+
+## [0.3.21] 2021/09/29
+
+## [0.3.20] 2021/09/29
+* Upgraded `heroku/jvm-function-invoker` to `0.5.3`
+
+## [0.3.19] 2021/09/29
+
+## [0.3.18] 2021/09/15
+* Upgraded `heroku/jvm` to `0.1.8`
+
+## [0.3.17] 2021/08/31
+* Upgraded `heroku/jvm-function-invoker` to `0.5.2`
 
 ## [0.3.16] 2021/08/25
 * Upgraded `heroku/jvm-function-invoker` to `0.5.1`