+++ shimmed-buildpacks/clojure/CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 * Fixed `licenses` in `buildpack.toml`
 
+### Changed
+* Switch to BSD 3-Clause License
+
 ## [0.0.87]
 This release is based on a classic Heroku buildpack via [cnb-shim](https://github.com/heroku/cnb-shim). The changelog
 for that release can be found here: https://github.com/heroku/heroku-buildpack-clojure/blob/main/CHANGELOG.md#v87